Move CompositeArrayBuffer to a new file
