+++ tfjs-core/src/io/weights_loader.ts
@@ -18,8 +18,8 @@
 import {env} from '../environment';
 
 import {NamedTensorMap} from '../tensor_types';
-import {TypedArray} from '../types';
 import * as util from '../util';
+import {CompositeArrayBuffer} from './composite_array_buffer';
 import {decodeWeights} from './io_utils';
 import {monitorPromisesProgress} from './progress';
 import {DTYPE_VALUE_SIZE_MAP, LoadOptions, WeightsManifestConfig, WeightsManifestEntry} from './types';
@@ -36,27 +36,27 @@ import {DTYPE_VALUE_SIZE_MAP, LoadOptions, WeightsManifestConfig, WeightsManifes
  *   length as `fetchURLs`.
  */
 export async function loadWeightsAsArrayBuffer(
-    fetchURLs: string[], loadOptions?: LoadOptions): Promise<ArrayBuffer[]> {
+  fetchURLs: string[], loadOptions?: LoadOptions): Promise<ArrayBuffer[]> {
   if (loadOptions == null) {
     loadOptions = {};
   }
 
   const fetchFunc = loadOptions.fetchFunc == null ? env().platform.fetch :
-                                                    loadOptions.fetchFunc;
+    loadOptions.fetchFunc;
 
   // Create the requests for all of the weights in parallel.
   const requests = fetchURLs.map(
-      fetchURL =>
-          fetchFunc(fetchURL, loadOptions.requestInit, {isBinary: true}));
+    fetchURL =>
+      fetchFunc(fetchURL, loadOptions.requestInit, { isBinary: true }));
 
   const fetchStartFraction = 0;
   const fetchEndFraction = 0.5;
 
   const responses = loadOptions.onProgress == null ?
-      await Promise.all(requests) :
-      await monitorPromisesProgress(
-          requests, loadOptions.onProgress, fetchStartFraction,
-          fetchEndFraction);
+    await Promise.all(requests) :
+    await monitorPromisesProgress(
+      requests, loadOptions.onProgress, fetchStartFraction,
+      fetchEndFraction);
 
   const bufferPromises = responses.map(response => response.arrayBuffer());
 
@@ -64,10 +64,10 @@ export async function loadWeightsAsArrayBuffer(
   const bufferEndFraction = 1;
 
   const buffers = loadOptions.onProgress == null ?
-      await Promise.all(bufferPromises) :
-      await monitorPromisesProgress(
-          bufferPromises, loadOptions.onProgress, bufferStartFraction,
-          bufferEndFraction);
+    await Promise.all(bufferPromises) :
+    await monitorPromisesProgress(
+      bufferPromises, loadOptions.onProgress, bufferStartFraction,
+      bufferEndFraction);
   return buffers;
 }
 
@@ -81,9 +81,9 @@ export async function loadWeightsAsArrayBuffer(
  * @param weightNames The names of the weights to be fetched.
  */
 export async function loadWeights(
-    manifest: WeightsManifestConfig, filePathPrefix = '',
-    weightNames?: string[],
-    requestInit?: RequestInit): Promise<NamedTensorMap> {
+  manifest: WeightsManifestConfig, filePathPrefix = '',
+  weightNames?: string[],
+  requestInit?: RequestInit): Promise<NamedTensorMap> {
   // TODO(nsthorat): Groups are currently fetched atomically. If you need a
   // single weight from a group, the whole group will be fetched. At a future
   // date, we should support fetching only the individual shards within a
@@ -91,7 +91,7 @@ export async function loadWeights(
   // TODO(cais): Use `decodeWeights` for implementation.
 
   const fetchWeights = (fetchUrls: string[]) =>
-      loadWeightsAsArrayBuffer(fetchUrls, {requestInit});
+    loadWeightsAsArrayBuffer(fetchUrls, { requestInit });
   const loadWeights = weightsLoaderFactory(fetchWeights);
 
   return loadWeights(manifest, filePathPrefix, weightNames);
@@ -122,12 +122,12 @@ export async function loadWeights(
  * @returns Weight loading function.
  */
 export function weightsLoaderFactory(
-    fetchWeightsFunction: (fetchUrls: string[]) => Promise<ArrayBuffer[]>):
-    (manifest: WeightsManifestConfig, filePathPrefix?: string,
-     weightNames?: string[]) => Promise<NamedTensorMap> {
-  return async(
-             manifest: WeightsManifestConfig, filePathPrefix = '',
-             weightNames?: string[]): Promise<NamedTensorMap> => {
+  fetchWeightsFunction: (fetchUrls: string[]) => Promise<ArrayBuffer[]>):
+  (manifest: WeightsManifestConfig, filePathPrefix?: string,
+    weightNames?: string[]) => Promise<NamedTensorMap> {
+  return async (
+    manifest: WeightsManifestConfig, filePathPrefix = '',
+    weightNames?: string[]): Promise<NamedTensorMap> => {
     // Collect all the groups, weights, and their relative offsets to be
     // fetched.
     const groupIndicesToFetchMap = manifest.map(() => false);
@@ -138,17 +138,17 @@ export function weightsLoaderFactory(
       }>
     } = {};
     const weightsFound =
-        weightNames != null ? weightNames.map(() => false) : [];
+      weightNames != null ? weightNames.map(() => false) : [];
     const allManifestWeightNames: string[] = [];
     manifest.forEach((manifestGroupConfig, groupIndex) => {
       let groupOffset = 0;
       manifestGroupConfig.weights.forEach(weightsEntry => {
         const rawDtype = ('quantization' in weightsEntry) ?
-            weightsEntry.quantization.dtype :
-            weightsEntry.dtype;
+          weightsEntry.quantization.dtype :
+          weightsEntry.dtype;
 
         const weightsBytes = DTYPE_VALUE_SIZE_MAP[rawDtype] *
-            util.sizeFromShape(weightsEntry.shape);
+          util.sizeFromShape(weightsEntry.shape);
 
         const enqueueWeightsForFetchingFn = () => {
           groupIndicesToFetchMap[groupIndex] = true;
@@ -182,27 +182,27 @@ export function weightsLoaderFactory(
     if (!weightsFound.every(found => found)) {
       const weightsNotFound = weightNames.filter((_, i) => !weightsFound[i]);
       throw new Error(
-          `Could not find weights in manifest with names: ` +
-          `${weightsNotFound.join(', ')}. \n` +
-          `Manifest JSON has weights with names: ` +
-          `${allManifestWeightNames.join(', ')}.`);
+        `Could not find weights in manifest with names: ` +
+        `${weightsNotFound.join(', ')}. \n` +
+        `Manifest JSON has weights with names: ` +
+        `${allManifestWeightNames.join(', ')}.`);
     }
 
     // Convert the one-hot boolean groupId => shouldFetch map to a list of group
     // IDs.
     const groupIndicesToFetch =
-        groupIndicesToFetchMap.reduce((accumulator, shouldFetch, i) => {
-          if (shouldFetch) {
-            accumulator.push(i);
-          }
-          return accumulator;
-        }, []);
+      groupIndicesToFetchMap.reduce((accumulator, shouldFetch, i) => {
+        if (shouldFetch) {
+          accumulator.push(i);
+        }
+        return accumulator;
+      }, []);
 
     const fetchUrls: string[] = [];
     groupIndicesToFetch.forEach(i => {
       manifest[i].paths.forEach(filepath => {
         const fetchUrl = filePathPrefix +
-            (!filePathPrefix.endsWith('/') ? '/' : '') + filepath;
+          (!filePathPrefix.endsWith('/') ? '/' : '') + filepath;
         fetchUrls.push(fetchUrl);
       });
     });
@@ -220,10 +220,10 @@ export function weightsLoaderFactory(
 
       weightsEntries.forEach(weightsEntry => {
         const byteBuffer = weightsBuffer.slice(
-            weightsEntry.groupOffset,
-            weightsEntry.groupOffset + weightsEntry.sizeBytes);
+          weightsEntry.groupOffset,
+          weightsEntry.groupOffset + weightsEntry.sizeBytes);
         const nameToTensorMap =
-            decodeWeights(byteBuffer, [weightsEntry.manifestEntry]);
+          decodeWeights(byteBuffer, [weightsEntry.manifestEntry]);
         for (const name in nameToTensorMap) {
           weightsTensorMap[name] = nameToTensorMap[name];
         }
@@ -234,191 +234,4 @@ export function weightsLoaderFactory(
 
     return weightsTensorMap;
   };
-}
-
-type BufferShard = {
-  start: number,
-  end: number,
-  buffer: ArrayBuffer,
-};
-
-/**
- * Wraps a list of ArrayBuffers into a `slice()`-able object without allocating
- * a large ArrayBuffer.
- *
- * Allocating large ArrayBuffers (~2GB) can be unstable on Chrome. TFJS loads
- * its weights as a list of (usually) 4MB ArrayBuffers and then slices the
- * weight tensors out of them. For small models, it's safe to concatenate all
- * the weight buffers into a single ArrayBuffer and then slice the weight
- * tensors out of it, but for large models, a different approach is needed.
- */
-export class CompositeArrayBuffer {
-  private shards: BufferShard[] = [];
-  private previousShardIndex = 0;
-  private bufferUniformSize?: number;
-  public readonly byteLength: number;
-
-  constructor(buffers: ArrayBuffer | ArrayBuffer[] | TypedArray
-      | TypedArray[]) {
-    // Normalize the `buffers` input to be `ArrayBuffer[]`.
-    if (!(buffers instanceof Array)) {
-      buffers = [buffers];
-    }
-    buffers = buffers.map((bufferOrTypedArray) => {
-      if (util.isTypedArray(bufferOrTypedArray)) {
-        return bufferOrTypedArray.buffer;
-      }
-      return bufferOrTypedArray;
-    });
-
-    // Skip setting up shards if there are no buffers.
-    if (buffers.length === 0) {
-      return;
-    }
-
-    this.bufferUniformSize = buffers[0].byteLength;
-    let start = 0;
-
-    for (let i = 0; i < buffers.length; i++) {
-      const buffer = buffers[i];
-      // Check that all buffers except the last one have the same length.
-      if (i !== buffers.length - 1 &&
-        buffer.byteLength !== this.bufferUniformSize) {
-        // Unset the buffer uniform size, since the buffer sizes are not
-        // uniform.
-        this.bufferUniformSize = undefined;
-      }
-
-      // Create the shards, including their start and end points.
-      const end = start + buffer.byteLength;
-      this.shards.push({buffer, start, end});
-      start = end;
-    }
-
-    // Set the byteLenghth
-    if (this.shards.length === 0) {
-      this.byteLength = 0;
-    }
-    this.byteLength = this.shards[this.shards.length - 1].end;
-  }
-
-  slice(start = 0, end = this.byteLength): ArrayBuffer {
-    // NaN is treated as zero for slicing. This matches ArrayBuffer's behavior.
-    start = isNaN(Number(start)) ? 0 : start;
-    end = isNaN(Number(end)) ? 0 : end;
-
-    // Fix the bounds to within the array.
-    start = Math.max(0, start);
-    end = Math.min(this.byteLength, end);
-    if (end <= start) {
-      return new ArrayBuffer(0);
-    }
-
-    const startShardIndex = this.findShardForByte(start);
-    if (startShardIndex === -1) {
-      // This should not happen since the start and end indices are always
-      // within 0 and the composite array's length.
-      throw new Error(`Could not find start shard for byte ${start}`);
-    }
-
-    const size = end - start;
-    const outputBuffer = new ArrayBuffer(size);
-    const outputArray = new Uint8Array(outputBuffer);
-    let sliced = 0;
-    for (let i = startShardIndex; i < this.shards.length; i++) {
-      const shard = this.shards[i];
-
-      const globalStart = start + sliced;
-      const localStart = globalStart - shard.start;
-      const outputStart = sliced;
-
-      const globalEnd = Math.min(end, shard.end);
-      const localEnd = globalEnd - shard.start;
-
-      const outputSlice = new Uint8Array(shard.buffer.slice(localStart,
-                                                            localEnd));
-      outputArray.set(outputSlice, outputStart);
-      sliced += outputSlice.length;
-
-      if (end < shard.end) {
-        break;
-      }
-    }
-    return outputBuffer;
-  }
-
-  /**
-   * Get the index of the shard that contains the byte at `byteIndex`.
-   */
-  private findShardForByte(byteIndex: number): number {
-    if (this.shards.length === 0 || byteIndex < 0 ||
-      byteIndex >= this.byteLength) {
-      return -1;
-    }
-
-    // If the buffers have a uniform size, compute the shard directly.
-    if (this.bufferUniformSize != null) {
-      this.previousShardIndex = Math.floor(byteIndex / this.bufferUniformSize);
-      return this.previousShardIndex;
-    }
-
-    // If the buffers don't have a uniform size, we need to search for the
-    // shard. That means we need a function to check where the byteIndex lies
-    // relative to a given shard.
-    function check(shard: BufferShard) {
-      if (byteIndex < shard.start) {
-        return -1;
-      }
-      if (byteIndex >= shard.end) {
-        return 1;
-      }
-      return 0;
-    }
-
-    // For efficiency, try the previous shard first.
-    if (check(this.shards[this.previousShardIndex]) === 0) {
-      return this.previousShardIndex;
-    }
-
-    // Otherwise, use a generic search function.
-    // This should almost never end up being used in practice since the weight
-    // entries should always be in order.
-    const index = search(this.shards, check);
-    if (index === -1) {
-      return -1;
-    }
-
-    this.previousShardIndex = index;
-    return this.previousShardIndex;
-  }
-}
-
-/**
- * Search for an element of a sorted array.
- *
- * @param sortedArray The sorted array to search
- * @param compare A function to compare the current value against the searched
- *     value. Return 0 on a match, negative if the searched value is less than
- *     the value passed to the function, and positive if the searched value is
- *     greater than the value passed to the function.
- * @returns The index of the element, or -1 if it's not in the array.
- */
-function search<T>(sortedArray: T[], compare: (t: T) => number): number {
-  // Binary search
-  let min = 0;
-  let max = sortedArray.length;
-
-  while (min <= max) {
-    const middle = Math.floor((max - min) / 2) + min;
-    const side = compare(sortedArray[middle]);
-
-    if (side === 0) {
-      return middle;
-    } else if (side < 0) {
-      max = middle;
-    } else {
-      min = middle + 1;
-    }
-  }
-  return -1;
 }
